Drop python2.7 and enhance the requirements and test
