+++ requirements_dev.txt
@@ -61,4 +61,4 @@ watchdog==0.9.0
 webencodings==0.5.1
 whichcraft==0.5.2
 wrapt==1.10.11
-yarl==1.3.0
+yarl>=1.2.0
