+++ setup.py
@@ -25,21 +25,19 @@ setup(
         'Intended Audience :: Developers',
         'License :: OSI Approved :: MIT License',
         'Natural Language :: English',
-        "Programming Language :: Python :: 2",
-        'Programming Language :: Python :: 2.7',
         'Programming Language :: Python :: 3',
         'Programming Language :: Python :: 3.4',
         'Programming Language :: Python :: 3.5',
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
     ],
-    description="Python Package (wrapper) for Quran Tafseer RESt APIs",
+    description="Python Package (wrapper) for Quran Tafseer REST APIs",
     install_requires=requirements,
     license="MIT license",
     long_description=readme + '\n\n' + history,
     include_package_data=True,
-    keywords='tafseer_package',
-    name='tafseer_package',
+    keywords='tafseer, quran, translation, tafsir',
+    name='pytafseer',
     packages=find_packages(include=['tafseer_package']),
     setup_requires=setup_requirements,
     test_suite='tests',
