+++ tests/test_tafseer_services.py
@@ -1,4 +1,4 @@
-from tafseer_package.services import QuranTafseer
+from tafseer_package.tafseer_package import QuranTafseer
 import pytest
 
 
@@ -6,8 +6,8 @@ import pytest
 def test_get_tafseer_books():
     tafseer_list = QuranTafseer.get_tafseer_books()
     assert tafseer_list is not None
-    assert list(tafseer_list[0].keys()) == ['id', 'name', 'language', 'author',
-                                            'book_name']
+    assert list(tafseer_list[0].keys()).sort() == ['id', 'name', 'language', 'author',
+                                                   'book_name'].sort()
 
 
 @pytest.mark.vcr()
@@ -21,18 +21,18 @@ def test_get_tafseer_books_with_lang():
 def test_get_tafseer_text_one_verse():
     book = QuranTafseer(1)
     tafseer_text = book.get_tafseer_text(1, 1)
-    assert list(tafseer_text.keys()) == ['tafseer_id', 'tafseer_name',
-                                         'ayah_url', 'ayah_number',
-                                         'text']
+    assert list(tafseer_text.keys()).sort() == ['tafseer_id', 'tafseer_name',
+                                                'ayah_url', 'ayah_number',
+                                                'text'].sort()
 
 
 @pytest.mark.vcr()
 def test_get_tafseer_text_one_verse_with_its_text():
     book = QuranTafseer(1)
     tafseer_text = book.get_tafseer_text(1, 1, with_verse_text=True)
-    assert list(tafseer_text.keys()) == ['tafseer_id', 'tafseer_name',
-                                         'ayah_url', 'ayah_number',
-                                         'text', 'verse_text']
+    assert list(tafseer_text.keys()).sort() == ['tafseer_id', 'tafseer_name',
+                                                'ayah_url', 'ayah_number',
+                                                'text', 'verse_text'].sort()
 
 
 @pytest.mark.vcr()
@@ -46,6 +46,6 @@ def test_get_tafseer_text_range_verse():
 def test_get_tafseer_text_range_verse_with_text():
     book = QuranTafseer(1)
     tafseer_text = book.get_tafseer_text_range(1, 1, 2, with_verse_text=True)
-    assert list(tafseer_text[0].keys()) == ['tafseer_id', 'tafseer_name',
-                                            'ayah_url', 'ayah_number',
-                                            'text', 'verse_text']
+    assert list(tafseer_text[0].keys()).sort() == ['tafseer_id', 'tafseer_name',
+                                                   'ayah_url', 'ayah_number',
+                                                   'text', 'verse_text'].sort()
